requirements.txt
Django~=2.2.13
django-cors-headers==3.1.1
django-filter==2.3.0
django-graphql-jwt==0.3.1
django-username-email==2.3.0
djangorestframework~=3.10.3
djangorestframework-simplejwt==4.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
graphene~=2.1.8
graphene-django==2.11.0
graphql-core==2.3.2
graphql-relay==2.0.1
pytz~=2019.3
gunicorn
psycopg2
